Move buff criteria to NBT
Rather than buffing based on whether the onering is in the player inventory the ring is now activated by sneak right clicking. This then adds an 'isActive' tag to the player NBT data. This is then checked during the player tick rather than checking the inventory,

In addition the ring activate is now used to keep track of whether the buff has been added to a player. The ring can only buff a player if it has not already done so. This prevents a single ring buffing many players. Sneak right clicking with a buffed player will remove the buff from the player and reset the ring such that it can be used again.

Buffs which need only be applied or removed once are now done on the right click event rather than done during every player tick.