Fix localisation and creative flight bug
Localisation is fixed by adding a pack.mcmeta file to the resource
folder.

Creative flight bug occurs when having the ring buff enabled, going
into creative and then survival. The change from creative to survival
removes the flight buff. To get around this activating a charged one
ring while already buffed will reenable flight without removing the
charge.

To help remove the buff from a player entirely an option has been added
to the base ring. Sneak right clicking will permanently remove the
buff.

In addition code relating to the activation via right clicking was
moved from the one ring to the event package. This allows for better
sharing for the new behaviour of the base ring and keeps the logic of
activating and deactivating the buff in one place.