Add item on activate buffs to the tick buff
The original intention was that certain buffs, like flight, only need
to be enabled when activating the ring. This unfortunately does not
work for cases where logging in or changing relms. In this case the
defaults are reinitalised, e.g. no flight, and without the tick the
on activate buffs are not enabled again.